return 404 if get by id returns null
